some security updated.

+++ modules/Cms/Http/Controllers/AcceptedProjectController.php
@@ -2,6 +2,7 @@
 
 namespace Modules\Cms\Http\Controllers;
 
+use App\Helpers\PermissionManager;
 use App\Http\Controllers\Controller;
 
 // requests...
@@ -70,6 +71,10 @@ class AcceptedProjectController extends Controller
             return redirect()->back();
         }
 
+        if (!PermissionManager::hasAcceptedPermission($project)) {
+            abort(404);
+        }
+
         $user = User::find(auth()->user()->id);
 
         if($user->hasRole('admin') || $user->hasRole('super_admin')) {
@@ -107,6 +112,9 @@ class AcceptedProjectController extends Controller
             // redirect back
             return redirect()->back();
         }
+        if (!PermissionManager::hasAcceptedPermission($project)) {
+            abort(404);
+        }
         // return view
         return view('cms::project.edit', compact('project'));
     }
@@ -128,6 +136,9 @@ class AcceptedProjectController extends Controller
             notifier()->error(__('admin/notifier.project_not_found'));
             // redirect back
             return redirect()->back();
+        }
+        if (!PermissionManager::hasAcceptedPermission($project)) {
+            abort(404);
         }
         // update project
         $project = $this->projectService->update($request->all(), $id);
@@ -161,6 +172,9 @@ class AcceptedProjectController extends Controller
             notifier()->error(__('admin/notifier.project_not_found'));
             // redirect back
             return redirect()->back();
+        }
+        if (!PermissionManager::hasAcceptedPermission($project)) {
+            abort(404);
         }
         // delete project
         if ($this->projectService->delete($id)) {
